+++ src/Presentation/SmartStore.Web/Controllers/ShoppingCartController.cs
@@ -283,7 +283,6 @@ namespace SmartStore.Web.Controllers
 				DisableWishlistButton = product.DisableWishlistButton
 			};
 
-			model.BasePrice = product.GetBasePriceInfo(_localizationService, _priceFormatter, _currencyService, _taxService, _priceCalculationService, customer, currency);
 			model.ProductUrl = _productUrlHelper.GetProductUrl(model.ProductSeName, sci);
 
 			if (item.BundleItem != null)
@@ -417,6 +416,8 @@ namespace SmartStore.Web.Controllers
 					model.Discount = _priceFormatter.FormatPrice(itemDiscount);
 				}
 
+                var basePriceAdjustment = (_priceCalculationService.GetFinalPrice(product, true) - _priceCalculationService.GetUnitPrice(sci, true)) * (-1);
+
                 model.BasePrice = product.GetBasePriceInfo(
                     _localizationService, 
                     _priceFormatter,
@@ -425,7 +426,7 @@ namespace SmartStore.Web.Controllers
                     _priceCalculationService,
 					customer,
                     currency,
-                    (product.Price - _priceCalculationService.GetUnitPrice(sci, true)) * (-1)
+                    basePriceAdjustment
                 );
 			}