+++ changelog.md
@@ -40,6 +40,7 @@
 * File upload of a checkout attribute was not stored on cart page.
 * Redirecting within checkout may have displayed an incorrect URL in the browser.
 * Server cannot modify cookies after HTTP headers have been sent.
+* Wrong base price on product and cart page when a special price is active.
 
 
 ## SmartStore.NET 3.1.5